LayoutTests/imported/w3c/web-platform-tests/requestidlecallback/basic-expected.txt

Basic requestIdleCallback Tests


PASS window.requestIdleCallback is defined
PASS window.cancelIdleCallback is defined
PASS window.requestIdleCallback() returns a number
PASS window.cancelIdleCallback() returns undefined
PASS requestIdleCallback schedules callbacks
PASS cancelIdleCallback cancels callbacks
